Added PlayerEffect enum.
Added proper microHID event support.

Removed GetHealth and SetHealth and replayed with a Health property and did the same with ArtificialHealth.

Added PlayerMedicalHealOverTimeEvent which fires when someone heals over time from medical items.

Added support for AHP in the PlayerHurtEvent.

Added Allow to PlayerShootEvent.

NOTE: `None of this has any backing code this is purely for discussion about what to add to the API`